# openapi.yaml
openapi: 3.0.3
info:
  title: Kanban System API
  description: 看板工作流管理系统 API
  version: 2.0.0
servers:
  - url: /api
    description: API Server
...
paths:
...
  /lists/{id}/cards:
    post:
      tags:
        - card
      summary: 创建卡片
      operationId: createCard
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCardRequest'
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Card'
  /cards/{id}:
    get:
      tags:
        - card
      summary: 获取卡片详情
      operationId: getCard
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Card'
    put:
      tags:
        - card
      summary: 更新卡片
      operationId: updateCard
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCardRequest'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Card'
    delete:
      tags:
        - card
      summary: 删除卡片
      operationId: deleteCard
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
  /cards/{id}/move:
    put:
      tags:
        - card
      summary: 移动卡片
      operationId: moveCard
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MoveCardRequest'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Card'
  /cards/{id}/complete:
    put:
      tags:
        - card
      summary: 完成卡片
      operationId: completeCard
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Card'
  /cards/{id}/assign:
    put:
      tags:
        - card
      summary: 指派卡片
      operationId: assignCard
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssignCardRequest'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Card'
  /cards/{id}/labels/{label_id}:
    post:
      tags:
        - card
      summary: 添加标签到卡片
      operationId: addLabelToCard
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: label_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
    delete:
      tags:
        - card
      summary: 从卡片移除标签
      operationId: removeLabelFromCard
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: label_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'

  # 检查清单接口
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        email:
          type: string
          format: email
        nickname:
          type: string
        avatar:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required:
        - id
        - username
        - email
...
    Card:
      type: object
      properties:
        id:
          type: integer
        list_id:
          type: integer
        swimlane_id:
          type: integer
          nullable: true
        title:
          type: string
        description:
          type: string
        position:
          type: integer
        due_date:
          type: string
          format: date-time
          nullable: true
        cover:
          type: string
        assignee_id:
          type: integer
          nullable: true
        assignee:
          $ref: '#/components/schemas/User'
        completed_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        labels:
          type: array
          items:
            $ref: '#/components/schemas/Label'
        comments:
          type: array
          items:
            $ref: '#/components/schemas/Comment'
        attachments:
          type: array
          items:
            $ref: '#/components/schemas/Attachment'
        checklist_items:
          type: array
          items:
            $ref: '#/components/schemas/ChecklistItem'
        checklist_progress:
          type: object
          properties:
            total:
              type: integer
            completed:
              type: integer
      required:
        - id
        - list_id
        - title
        - position
    ChecklistItem:
      type: object
      properties:
        id:
          type: integer
        card_id:
          type: integer
        content:
          type: string
        completed:
          type: boolean
        position:
          type: integer
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required:
        - id
        - card_id
        - content
        - completed
        - position
    Label:
      type: object
      properties:
        id:
          type: integer
        board_id:
          type: integer
        name:
          type: string
        color:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required:
        - id
        - board_id
        - name
        - color
    Comment:
      type: object
      properties:
        id:
          type: integer
        card_id:
          type: integer
        user_id:
          type: integer
        user:
          $ref: '#/components/schemas/User'
        content:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required:
        - id
        - card_id
        - user_id
        - content
    Attachment:
      type: object
      properties:
        id:
          type: integer
        card_id:
          type: integer
        file_name:
          type: string
        file_url:
          type: string
        file_size:
          type: integer
        mime_type:
          type: string
        created_at:
          type: string
          format: date-time
      required:
        - id
        - card_id
        - file_name
        - file_url
...
    CreateCardRequest:
      type: object
      properties:
        title:
          type: string
          maxLength: 200
        description:
          type: string
        swimlane_id:
          type: integer
          nullable: true
        due_date:
          type: string
          format: date-time
          nullable: true
        cover:
          type: string
        assignee_id:
          type: integer
          nullable: true
      required:
        - title
    UpdateCardRequest:
      type: object
      properties:
        title:
          type: string
          maxLength: 200
        description:
          type: string
        position:
          type: integer
        list_id:
          type: integer
        swimlane_id:
          type: integer
          nullable: true
        due_date:
          type: string
          format: date-time
          nullable: true
        cover:
          type: string
        assignee_id:
          type: integer
          nullable: true
    MoveCardRequest:
      type: object
      properties:
        list_id:
          type: integer
        position:
          type: integer
      required:
        - list_id
    AssignCardRequest:
      type: object
      properties:
        assignee_id:
          type: integer
          nullable: true
      required:
        - assignee_id
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message